+++ stroom-security/src/main/java/stroom/security/server/DBRealm.java
@@ -125,7 +125,7 @@ public class DBRealm extends AuthenticatingRealm {
         } catch (final AuthenticationException e) {
             throw e;
         } catch (final Exception e) {
-            LOGGER.error(e.getMessage(), e);
+            LOGGER.debug(e.getMessage(), e);
             throw new BadCredentialsException(e.getMessage());
         }
 
@@ -155,7 +155,7 @@ public class DBRealm extends AuthenticatingRealm {
         } catch (final AuthenticationException e) {
             throw e;
         } catch (final Exception e) {
-            LOGGER.error(e.getMessage(), e);
+            LOGGER.debug(e.getMessage(), e);
             throw new BadCredentialsException(e.getMessage());
         }